Move select range trigger to uiActions (#60168)
* move select range trigger to uiActions
