+++ src/plugins/embeddable/public/bootstrap.ts
@@ -23,14 +23,12 @@ import {
   contextMenuTrigger,
   createFilterAction,
   panelBadgeTrigger,
-  selectRangeTrigger,
   valueClickTrigger,
   EmbeddableVisTriggerContext,
   IEmbeddable,
   EmbeddableContext,
   APPLY_FILTER_TRIGGER,
   VALUE_CLICK_TRIGGER,
-  SELECT_RANGE_TRIGGER,
   CONTEXT_MENU_TRIGGER,
   PANEL_BADGE_TRIGGER,
   ACTION_ADD_PANEL,
@@ -44,7 +42,6 @@ import {
 
 declare module '../../ui_actions/public' {
   export interface TriggerContextMapping {
-    [SELECT_RANGE_TRIGGER]: EmbeddableVisTriggerContext;
     [VALUE_CLICK_TRIGGER]: EmbeddableVisTriggerContext;
     [APPLY_FILTER_TRIGGER]: {
       embeddable: IEmbeddable;
@@ -72,7 +69,6 @@ export const bootstrap = (uiActions: UiActionsSetup) => {
   uiActions.registerTrigger(contextMenuTrigger);
   uiActions.registerTrigger(applyFilterTrigger);
   uiActions.registerTrigger(panelBadgeTrigger);
-  uiActions.registerTrigger(selectRangeTrigger);
   uiActions.registerTrigger(valueClickTrigger);
 
   const actionApplyFilter = createFilterAction();
