+++ src/plugins/embeddable/public/index.ts
@@ -62,8 +62,6 @@ export {
   PanelNotFoundError,
   PanelState,
   PropertySpec,
-  SELECT_RANGE_TRIGGER,
-  selectRangeTrigger,
   VALUE_CLICK_TRIGGER,
   valueClickTrigger,
   ViewMode,
